refactored app.js to render the whole app, not just the login. deleted duplicate login.js and pages dir
