+++ client/src/components/appContainer.js
@@ -9,7 +9,7 @@ import IndividualRecipe from './pages/individualRecipe';
 import AddRecipe from './pages/addRecipe';
 
 export default function AppContainer() {
-  const [currentPage, setCurrentPage] = useState('Homepage');
+  const [currentPage, setCurrentPage] = useState('Login');
 
   const renderPage = () => {
     switch (currentPage) {
